chore: move component to new components rename files to `image`
