+++ src/new-components/image/image.stories.mdx
@@ -1,8 +1,8 @@
 import { Canvas, Meta, Story } from '@storybook/addon-docs';
-import { Image } from './image-content-block';
-import { Regular } from './image-content-block.stories';
+import { Image } from './image';
+import { Regular } from './image.stories';
 
-<Meta title="Components/Image" component={Image} />
+<Meta title="Components/Image" component={{ Image }} />
 
 # Image
